Refine collision to account for relative speed
A collision must occur along the direction of the colliding objects
relative velocities. For now, this is simply the Speed<> of the ball,
though it can be further improved in the future to account for
translation differences between gameloops for the generic polygon.

This solves the issue of a collision happening too deep into respective
polygons and being misattributed to the backside of the polygon.

A remaining issue is a ball travelling so fast that it jumps over the
polygon entirely, skipping collision detection.

+++ pong/ball.h
@@ -78,7 +78,7 @@ template <typename T> class Ball : public Collider<T>, public Renderable
         return m_placement.size();
     };
 
-    [[nodiscard]] const Speed<float>& getSpeed() const noexcept
+    [[nodiscard]] const Speed<double>& getSpeed() const noexcept
     {
         return m_speed;
     };
